pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kalshi-weather-bot"
version = "0.1.0"
description = "Research stack for weather-based Kalshi market modeling"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "requests>=2.32.0",
  "pydantic>=2.8.0",
  "pyyaml>=6.0.2",
  "duckdb>=1.1.0",
  "pandas>=2.2.0",
  "pyarrow>=17.0.0",
  "python-dotenv>=1.0.1",
  "typer>=0.12.3",
  "rich>=13.7.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "ruff>=0.6.0",
  "mypy>=1.11.0",
]

[project.scripts]
kwb = "kwb.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
